# src/handlers/orderInfo.ts
import { APIGatewayProxyEvent } from 'aws-lambda';
import { getOrderInfo } from '../order';
import { InvalidOrderId, UnauthorisedError } from '../throwError';
import { jsonResponse } from './response';

export const getOrderInfoHandler = async (event: APIGatewayProxyEvent) => {
  try {
    // get the session from the header
    const session = event.headers.session;
    if (!session) {
      return jsonResponse(400, { error: 'provided session is not valid'});
    }
    // check if the provided path contains an orderId
    if (!event.pathParameters) {
      return jsonResponse(400, { error: 'orderId is null.' });
    }
    // get the orderId from the route path
    const orderId = event.pathParameters.orderId;
    if (!orderId) {
      return jsonResponse(400, { error: 'orderId provided is not valid' });
    }

    // call the backend function
    const res = await getOrderInfo(session, orderId);

    return jsonResponse(200, res);
  } catch (e) {
    if (e instanceof InvalidOrderId) {
      return jsonResponse(400, { error: e.message });
    }
    if (e instanceof UnauthorisedError) {
      return jsonResponse(401, { error: e.message });
    }
    // internal server error, server doesnot know how to handle the error
    return jsonResponse(500, { error: 'INTERNAL SERVER ERROR' });
  }
};