# swagger.yaml
swagger: "2.0"
info:
  title: Order Creation Service
  description: Important operations required to manage order creation data effectively by letting the user to create, retrieve, list, update and cancel (delete) orders when needed. Please ensure to input your api key on the "Authorize" field.
  version: 1.0.0

host: "tx12cup9z0.execute-api.ap-southeast-2.amazonaws.com"
basePath: "/Sprint2Final"
schemes:
  - https

securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: x-api-key

security:
  - ApiKeyAuth: []

tags:
  - name: PHASE 1
  - name: Users
    description: Deals with all the user related operations which includes user register, login, update and logout
  - name: Orders
    description: Deals with all the order related operations which includes creating, viewing, listing, updating and cancelling any order
  - name: Other

definitions:
  userRegister:
    type: object
    properties:
      firstName:
        type: string
        example: Anna
      lastName: 
        type: string
        example: Smith
      email:
        type: string
        example: anna123@gmail.com
      telephone:
        type: string
        example: "0412345678"
      password: 
        type: string
        example: h3llow0rld

  error400:
    type: object
    properties:
      error:
        type: string
        example: "INVALID_EMAIL"

  error401:
    type: object
    properties:
      error:
        type: string
        example: "Unauthorised access"

  session:
    type: object
    properties:
      session: 
        type: string
        example: "eyJhbGciOiJIUzI1NiIsInR5c..."
  
  userLogIn:
    type: object
    properties:
      email:
        type: string
        example: anna123@gmail.com
      password: 
        type: string
        example: h3llow0rld
  
  emptyObj:
    type: object
    properties: {}
  
  userUpdate:
    type: object
    properties:
      email:
        type: string
        example: anna123@gmail.com
      firstName:
        type: string
        example: Brianna
      lastName:
        type: string
        example: Lee
      telephone:
        type: string
        example: "0412345678"
  
  orderId:
    type: object
    properties:
      orderId:
        type: string
        example: "550e8400-e29b-41d4-a716-446655440000"
  
  orderItem: 
    type: object
    properties:
      name: 
        type: string
        example: Chocolate Bar
      description:
        type: string
        example: Milk Chocolate
      unitPrice:
        type: number
        example: 10
      quantity:
        type: number
        example: 1
  
  orderCreate:
    type: object
    properties:
      currency:
        type: string
        example: AUD
      deliveryAddress:
        type: string
        example: 123 Street Name, Kensington
      user:
        type: object
        properties:
          firstName:
            type: string
            example: Anna
          lastName:
            type: string
            example: Smith
          telephone:
            type: string
            example: "0412345678"
          email: 
            type: string
            example: anna123@gmail.com
      reqDeliveryPeriod:
        type: object
        properties: 
          startDateTime: 
            type: number
            example: 1773100800000
          endDateTime:
            type: number
            example: 1773619200000
      items:
        type: array
        items:
            $ref: '#/definitions/orderItem'
  
  orderView:
    type: object
    properties:
      orderId:
        type: string
        example: "550e8400-e29b-41d4-a716-446655440000"
      status:
        type: string
        example: OPEN
      issuedDate:
        type: string
        example: "2026-03-16"
      issuedTime:
        type: string
        example: "16:10:05"
      currency:
        type: string
        example: AUD
      taxExclusive:
        type: number
        example: 100
      taxInclusive:
        type: number
        example: 110
      finalPrice:
        type: number
        example: 110
      address:
        type: string
        example: 123 Street Name, Kensington
      userDetails:
        type: object
        properties: 
          firstName: { type: string, example: "Anna" }
          lastName: { type: string, example: "Smith" }
          telephone: { type: string, example: "0412345678" }
          email: { type: string, example: "anna123@gmail.com" }
      deliveryDetails:
        type: object
        properties: 
          startDateTime: { type: number, example: 1773100800000 }
          endDateTime: { type: number, example: 1773619200000 }
      items:
        type: array
        items:
            $ref: '#/definitions/orderItem'
  
  orderCancel:
    type: object
    properties:
      reason:
        type: string
        example: Change of Mind

  orderUpdate:
    type: object
    properties:
      deliveryAddress:
        type: string
        example: 123 Street Name, Kensington
      reqDeliveryPeriod:
        type: object
        properties: 
          startDateTime: { type: number, example: 1773100800000 }
          endDateTime: { type: number, example: 1773619200000 }
      status:
        type: string
        example: "UPDATED"
  
  orderList:
    type: object
    properties:
      orders:
        type: array
        items:
          type: object
          properties:
            orderId: { type: string, example: "550e8400-e29b-41d4-a716-446655440000" }
            status: { type: string, example: "OPEN" }
            issuedDate: { type: string, example: "2026-03-16" }
            finalPrice: { type: number, example: 110 }
            currency: { type: string, example: "AUD" }

paths:
  /v0/user/register:
    post:
      tags: [Users]
      summary: Register a new user 
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userRegister'
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/session'
        400:
          description: |
            **INVALID INPUT**
            - INVALID_FIRST_NAME
            - INVALID_LAST_NAME
            - INVALID_EMAIL
            - INVALID_PASSWORD_FORMAT
          schema:
            $ref: '#/definitions/error400'

  /v0/user/login:
    post:
      tags: [Users]
      summary: Login an existing user
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userLogIn'
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/session'
        400:
          description: |
            **INVALID USER DETAILS**
            - INVALID_EMAIL
            - INCORRECT_PASSWORD
          schema:
            $ref: '#/definitions/error400'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'
  
  /v0/user/details:
    put:
      tags: [Users]
      summary: Update user details
      parameters:
        - in: header
          name: session
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/userUpdate'
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/emptyObj'
        400:
          description: "**INVALID INPUT**"
          schema:
            $ref: '#/definitions/error400'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'
  
  /v0/user/logout:
    post:
      tags: [Users]
      summary: Logs out an existing user
      parameters:
        - in: header
          name: session
          required: true
          type: string
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/emptyObj'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'
  
  /v0/order:
    post:
      tags: [Orders]
      summary: Creates an Order
      parameters:
        - in: header
          name: session
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/orderCreate'
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/orderId'
        400:
          description: |
             "**INVALID INPUT**"
             - INVALID_EMAIL
             - INVALID_MOBILE_NO
             - INVALID_ITEM
          schema:
            $ref: '#/definitions/error400'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'

  /v0/order/{orderId}:
    get:
      tags: [Orders]
      summary: View order information 
      parameters:
        - in: path
          name: orderId
          required: true
          type: string
        - in: header
          name: session
          required: true
          type: string
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/orderView'
        400:
          description: "**INVALID INPUT**"
          schema:
            $ref: '#/definitions/error400'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'

    delete:
      tags: [Orders]
      summary: Cancel an order 
      parameters:
        - in: path
          name: orderId
          required: true
          type: string
        - in: header
          name: session
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/orderCancel'
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/orderCancel'
        400:
          description: "**INVALID INPUT**"
          schema:
            $ref: '#/definitions/error400'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'
            
    put:
      tags: [Orders]
      summary: Update an order 
      parameters:
        - in: path
          name: orderId
          required: true
          type: string
        - in: header
          name: session
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/orderUpdate'
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/emptyObj'
        400:
          description: "**INVALID INPUT**"
          schema:
            $ref: '#/definitions/error400'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'

  /v0/order/list:
    get:
      tags: [Orders]
      summary: Get the list of all orders
      parameters:
        - in: header
          name: session
          required: true
          type: string
      responses:
        200:
          description: "**SUCCESS**"
          schema:
            $ref: '#/definitions/orderList'
        401:
          description: "**UNAUTHORISED**"
          schema: 
            $ref: '#/definitions/error401'